오류 못잡음
합치기 힘든상태

+++ backend/src/main/java/com/example/backend/hotel/dto/HotelSearchResponseDto.java
@@ -12,12 +12,12 @@ public class HotelSearchResponseDto {
     private Long hotelId;
     private String name;
     private String address;
-    private BigDecimal grade;
+    private Double grade;
     private String imageUrl;
-    private BigDecimal minPrice;
+    private Double minPrice;
 
     // JPQL에서 new 생성자를 사용하기 위한 생성자
-    public HotelSearchResponseDto(Long hotelId, String name, String address, BigDecimal grade, String imageUrl, BigDecimal minPrice) {
+    public HotelSearchResponseDto(Long hotelId, String name, String address, Double grade, String imageUrl, Double minPrice) {
         this.hotelId = hotelId;
         this.name = name;
         this.address = address;
